バリデータとバリデーション実行をファイルを分けて記述
